deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy
plotly
pillow
tensorflow-cpu
scikit-learn==1.2.1
imblearn
feature_engine
